Move various I/O related metrics files into subdir
Bug: 550013241
