+++ base/cvd/cuttlefish/host/commands/metrics_transmitter/metrics_flags.h
@@ -18,7 +18,7 @@
 
 #include <string>
 
-#include "cuttlefish/host/libs/metrics/metrics_environment.h"
+#include "cuttlefish/host/libs/metrics/io/metrics_environment.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
