+++ base/cvd/cuttlefish/host/commands/metrics_transmitter/metrics_transmission.cc
@@ -19,7 +19,7 @@
 #include <memory>
 #include <string>
 
-#include "cuttlefish/host/libs/metrics/metrics_environment.h"
+#include "cuttlefish/host/libs/metrics/io/metrics_environment.h"
 #include "cuttlefish/host/libs/web/http_client/curl_global_init.h"
 #include "cuttlefish/host/libs/web/http_client/curl_http_client.h"
 #include "cuttlefish/host/libs/web/http_client/http_client.h"
