+++ base/cvd/cuttlefish/host/commands/start/BUILD.bazel
@@ -32,7 +32,7 @@ cf_cc_binary(
         "//cuttlefish/host/libs/config:host_tools_version",
         "//cuttlefish/host/libs/config:instance_nums",
         "//cuttlefish/host/libs/log_names",
-        "//cuttlefish/host/libs/metrics:notification",
+        "//cuttlefish/host/libs/metrics/io:notification",
         "//cuttlefish/host/libs/vm_manager",
         "//cuttlefish/posix:readlink",
         "//cuttlefish/posix:symlink",
