+++ base/cvd/cuttlefish/host/libs/metrics/BUILD.bazel
@@ -15,24 +15,15 @@ cf_cc_library(
         "//cuttlefish/files:file_exists",
         "//cuttlefish/host/commands/cvd/instances",
         "//cuttlefish/host/libs/metrics:metrics_orchestration",
-        "//cuttlefish/host/libs/metrics:notification",
         "//cuttlefish/host/libs/metrics/conversion:metrics_data",
         "//cuttlefish/host/libs/metrics/guest:device_event_type",
         "//cuttlefish/host/libs/metrics/guest:guest_metrics",
         "//cuttlefish/host/libs/metrics/guest:parsed_flags",
+        "//cuttlefish/host/libs/metrics/io:notification",
         "//cuttlefish/result",
         "@abseil-cpp//absl/log",
         "@abseil-cpp//absl/strings",
         "@fmt",
-    ],
-)
-
-cf_cc_library(
-    name = "enabled",
-    srcs = ["enabled.cc"],
-    hdrs = ["enabled.h"],
-    deps = [
-        "//cuttlefish/files:file_exists",
     ],
 )
 
@@ -59,18 +50,12 @@ cf_cc_library(
         "//cuttlefish/host/commands/cvd/fetch:fetch_cvd_parser",
         "//cuttlefish/host/libs/metrics:fetch_metrics",
         "//cuttlefish/host/libs/metrics:metrics_orchestration",
-        "//cuttlefish/host/libs/metrics:notification",
         "//cuttlefish/host/libs/metrics/conversion:metrics_data",
+        "//cuttlefish/host/libs/metrics/io:notification",
         "//cuttlefish/result",
         "@abseil-cpp//absl/log",
         "@fmt",
     ],
-)
-
-cf_cc_library(
-    name = "metrics_environment",
-    srcs = ["metrics_environment.cc"],
-    hdrs = ["metrics_environment.h"],
 )
 
 cf_cc_library(
@@ -82,69 +67,14 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:tee_logging",
         "//cuttlefish/host/commands/cvd/version",
         "//cuttlefish/host/libs/log_names",
-        "//cuttlefish/host/libs/metrics:enabled",
-        "//cuttlefish/host/libs/metrics:metrics_transmitter",
-        "//cuttlefish/host/libs/metrics:metrics_writer",
-        "//cuttlefish/host/libs/metrics:session_id",
-        "//cuttlefish/host/libs/metrics/host:host_metrics",
         "//cuttlefish/host/libs/metrics/conversion:metrics_data",
+        "//cuttlefish/host/libs/metrics/host:host_metrics",
+        "//cuttlefish/host/libs/metrics/io:enabled",
+        "//cuttlefish/host/libs/metrics/io:metrics_transmitter",
+        "//cuttlefish/host/libs/metrics/io:metrics_writer",
+        "//cuttlefish/host/libs/metrics/io:session_id",
         "//cuttlefish/result",
         "//external_proto:cf_log_cc_proto",
-        "@fmt",
-    ],
-)
-
-cf_cc_library(
-    name = "metrics_transmitter",
-    srcs = ["metrics_transmitter.cc"],
-    hdrs = ["metrics_transmitter.h"],
-    deps = [
-        "//cuttlefish/common/libs/utils:base64",
-        "//cuttlefish/host/libs/metrics:metrics_environment",
-        "//cuttlefish/process:command",
-        "//cuttlefish/process:managed_stdio",
-        "//cuttlefish/result",
-        "//external_proto:cf_log_cc_proto",
-    ],
-)
-
-cf_cc_library(
-    name = "metrics_writer",
-    srcs = ["metrics_writer.cc"],
-    hdrs = ["metrics_writer.h"],
-    deps = [
-        "//cuttlefish/common/libs/utils:files",
-        "//cuttlefish/common/libs/utils:random",
-        "//cuttlefish/result",
-        "//external_proto:cf_log_cc_proto",
-        "@fmt",
-        "@protobuf",
-    ],
-)
-
-cf_cc_library(
-    name = "notification",
-    srcs = ["notification.cc"],
-    hdrs = ["notification.h"],
-    deps = [
-        "//cuttlefish/files:file_exists",
-        "//cuttlefish/host/libs/directories",
-        "//cuttlefish/host/libs/metrics:enabled",
-        "//cuttlefish/result",
-        "@abseil-cpp//absl/log",
-        "@fmt",
-    ],
-)
-
-cf_cc_library(
-    name = "session_id",
-    srcs = ["session_id.cc"],
-    hdrs = ["session_id.h"],
-    deps = [
-        "//cuttlefish/common/libs/utils:files",
-        "//cuttlefish/common/libs/utils:uuid",
-        "//cuttlefish/files:file_exists",
-        "//cuttlefish/result",
         "@fmt",
     ],
 )
