+++ base/cvd/cuttlefish/host/libs/metrics/device_metrics_orchestration.cc
@@ -31,8 +31,8 @@
 #include "cuttlefish/host/libs/metrics/guest/device_event_type.h"
 #include "cuttlefish/host/libs/metrics/guest/guest_metrics.h"
 #include "cuttlefish/host/libs/metrics/guest/parsed_flags.h"
+#include "cuttlefish/host/libs/metrics/io/notification.h"
 #include "cuttlefish/host/libs/metrics/metrics_orchestration.h"
-#include "cuttlefish/host/libs/metrics/notification.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
