+++ base/cvd/cuttlefish/host/libs/metrics/io/BUILD.bazel
@@ -1,0 +1,77 @@
+load("//cuttlefish/bazel:rules.bzl", "cf_build_test", "cf_cc_library")
+
+package(
+    default_visibility = ["//:android_cuttlefish"],
+)
+
+cf_build_test(name = "io_build_test")
+
+cf_cc_library(
+    name = "enabled",
+    srcs = ["enabled.cc"],
+    hdrs = ["enabled.h"],
+    deps = [
+        "//cuttlefish/files:file_exists",
+    ],
+)
+
+cf_cc_library(
+    name = "metrics_environment",
+    srcs = ["metrics_environment.cc"],
+    hdrs = ["metrics_environment.h"],
+)
+
+cf_cc_library(
+    name = "metrics_transmitter",
+    srcs = ["metrics_transmitter.cc"],
+    hdrs = ["metrics_transmitter.h"],
+    deps = [
+        "//cuttlefish/common/libs/utils:base64",
+        "//cuttlefish/host/libs/metrics/io:metrics_environment",
+        "//cuttlefish/process:command",
+        "//cuttlefish/process:managed_stdio",
+        "//cuttlefish/result",
+        "//external_proto:cf_log_cc_proto",
+    ],
+)
+
+cf_cc_library(
+    name = "metrics_writer",
+    srcs = ["metrics_writer.cc"],
+    hdrs = ["metrics_writer.h"],
+    deps = [
+        "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/common/libs/utils:random",
+        "//cuttlefish/result",
+        "//external_proto:cf_log_cc_proto",
+        "@fmt",
+        "@protobuf",
+    ],
+)
+
+cf_cc_library(
+    name = "notification",
+    srcs = ["notification.cc"],
+    hdrs = ["notification.h"],
+    deps = [
+        "//cuttlefish/files:file_exists",
+        "//cuttlefish/host/libs/directories",
+        "//cuttlefish/host/libs/metrics/io:enabled",
+        "//cuttlefish/result",
+        "@abseil-cpp//absl/log",
+        "@fmt",
+    ],
+)
+
+cf_cc_library(
+    name = "session_id",
+    srcs = ["session_id.cc"],
+    hdrs = ["session_id.h"],
+    deps = [
+        "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/common/libs/utils:uuid",
+        "//cuttlefish/files:file_exists",
+        "//cuttlefish/result",
+        "@fmt",
+    ],
+)
