+++ base/cvd/cuttlefish/host/libs/metrics/io/enabled.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2025 The Android Open Source Project
+ * Copyright (C) 2026 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,21 +14,12 @@
  * limitations under the License.
  */
 
-#include "cuttlefish/host/libs/metrics/metrics_environment.h"
+#include "cuttlefish/host/libs/metrics/io/enabled.h"
 
-#include <string>
+#include "cuttlefish/files/file_exists.h"
 
 namespace cuttlefish {
 
-std::string EnvironmentToString(ClearcutEnvironment environment) {
-  switch (environment) {
-    case ClearcutEnvironment::Local:
-      return std::string(kClearcutLocal);
-    case ClearcutEnvironment::Staging:
-      return std::string(kClearcutStaging);
-    case ClearcutEnvironment::Production:
-      return std::string(kClearcutProduction);
-  }
-}
+bool AreMetricsEnabled() { return FileExists(kTransmitterPath); }
 
 }  // namespace cuttlefish
