+++ base/cvd/cuttlefish/host/libs/metrics/io/metrics_environment.h
