+++ base/cvd/cuttlefish/host/libs/metrics/io/metrics_transmitter.cc
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-#include "cuttlefish/host/libs/metrics/metrics_transmitter.h"
+#include "cuttlefish/host/libs/metrics/io/metrics_transmitter.h"
 
 #include <string>
 #include <utility>
 
 #include "cuttlefish/common/libs/utils/base64.h"
-#include "cuttlefish/host/libs/metrics/metrics_environment.h"
+#include "cuttlefish/host/libs/metrics/io/metrics_environment.h"
 #include "cuttlefish/process/command.h"
 #include "cuttlefish/process/managed_stdio.h"
 #include "cuttlefish/result/result.h"
