+++ base/cvd/cuttlefish/host/libs/metrics/io/session_id.h
