+++ base/cvd/cuttlefish/host/libs/metrics/metrics_orchestration.cc
@@ -29,11 +29,11 @@
 #include "cuttlefish/host/commands/cvd/version/version.h"
 #include "cuttlefish/host/libs/log_names/log_names.h"
 #include "cuttlefish/host/libs/metrics/conversion/metrics_data.h"
-#include "cuttlefish/host/libs/metrics/enabled.h"
 #include "cuttlefish/host/libs/metrics/host/host_metrics.h"
-#include "cuttlefish/host/libs/metrics/metrics_transmitter.h"
-#include "cuttlefish/host/libs/metrics/metrics_writer.h"
-#include "cuttlefish/host/libs/metrics/session_id.h"
+#include "cuttlefish/host/libs/metrics/io/enabled.h"
+#include "cuttlefish/host/libs/metrics/io/metrics_transmitter.h"
+#include "cuttlefish/host/libs/metrics/io/metrics_writer.h"
+#include "cuttlefish/host/libs/metrics/io/session_id.h"
 #include "cuttlefish/result/result.h"
 #include "external_proto/cf_log.pb.h"
